fix: move useEthereumName to individual file
